refactor(cru): 将 PLL 参数查找函数模块化并优化常量定义
- 将 find_pll_params 从 Cru 方法移为 pll 模块独立函数
- 在 variants 模块统一导出 MHZ 常量，移除重复定义
- 优化 PLL 参数查找逻辑的代码组织和可维护性
