+++ src/variants/mod.rs
@@ -1,3 +1,3 @@
 pub mod rk3588;
 
-const MHZ: u64 = 1_000_000;
+pub(crate) const MHZ: u64 = 1_000_000;
