+++ src/variants/rk3588/cru/consts.rs
@@ -8,7 +8,8 @@
 // 频率常量
 // ============================================================================
 
-pub const MHZ: u64 = 1_000_000;
+pub(crate) use crate::MHZ;
+
 pub const KHZ: u32 = 1_000;
 pub const OSC_HZ: u64 = 24 * MHZ;
